Generate fantasy schedules

This function generates random head to head schedules for a given number of seasons, teams, and weeks.

Usage

ffs_build_schedules(
  n_teams = NULL,
  n_seasons = 100,
  n_weeks = 14,
  franchises = NULL,
  seed = NULL
)

Arguments

n_teams

number of teams in simulation

n_seasons

number of seasons to simulate, default = 100

n_weeks

number of weeks per season, default = 14

franchises

optional: a dataframe of franchises as created by ffs_franchises() - overrides the n_teams argument and will attach actual franchise IDs to the schedule output.

seed

an integer to control reproducibility

Value

a dataframe of schedules

Details

It starts with the circle method for round robin scheduling, grows or shrinks the schedule to match the required number of weeks, and then shuffles both the order that teams are assigned in and the order that weeks are generated. This doesn't "guarantee" unique schedules, but there are n_teams! x n_weeks! permutations of the schedule so it's very very likely that the schedules are unique (3x10^18 possible schedules for a 12 team league playing 13 weeks).

See also

Examples

# \donttest{
ffs_build_schedules(n_teams = 12, n_seasons = 1, n_weeks = 14)
#>      season  week franchise_id opponent_id
#>       <int> <int>        <int>       <int>
#>   1:      1     5            6          10
#>   2:      1     5           11           9
#>   3:      1     5            2           3
#>   4:      1     5            8           7
#>   5:      1     5            5           4
#>  ---                                      
#> 164:      1     9            4          10
#> 165:      1     9            7           9
#> 166:      1     9            3           6
#> 167:      1     9            9           7
#> 168:      1     9           10           4
# }