Generate Projections

Runs the bootstrapped resampling of player week outcomes on the latest rankings and rosters for a given number of seasons and weeks per season.

ffs_generate_projections(
  adp_outcomes,
  latest_rankings,
  n_seasons = 100,
  weeks = 1:14,
  rosters = NULL
)

Arguments

adp_outcomes

a dataframe of adp-based weekly outcomes, as created by ffs_adp_outcomes()

latest_rankings

a dataframe of rankings, as created by ffs_latest_rankings()

n_seasons

number of seasons, default is 100

weeks

a numeric vector of weeks to simulate, defaults to 1:14

rosters

a dataframe of rosters, as created by ffs_rosters() - optional, reduces computation to just rostered players

Value

a dataframe of weekly scores for each player in the simulation, approximately of length n_seasons x n_weeks x latest_rankings

See also

Examples

# \donttest{
# cached examples
adp_outcomes <- .ffs_cache("adp_outcomes.rds")
latest_rankings <- .ffs_cache("latest_rankings.rds")

ffs_generate_projections(adp_outcomes, latest_rankings)
#>         season fantasypros_id             player pos team bye    ecr    sd rank
#>      1:      2          16413 Patrick Mahomes II  QB  KCC  12   1.32  0.69    1
#>      2:      2          16413 Patrick Mahomes II  QB  KCC  12   1.32  0.69    1
#>      3:      2          16413 Patrick Mahomes II  QB  KCC  12   1.32  0.69    1
#>      4:      2          16413 Patrick Mahomes II  QB  KCC  12   1.32  0.69    1
#>      5:      2          16413 Patrick Mahomes II  QB  KCC  12   1.32  0.69    1
#>     ---                                                                        
#> 994374:     52          19534        Tarik Black  WR  IND  14 239.50 20.50  243
#> 994375:     52          19534        Tarik Black  WR  IND  14 239.50 20.50  243
#> 994376:     52          19534        Tarik Black  WR  IND  14 239.50 20.50  243
#> 994377:     52          19534        Tarik Black  WR  IND  14 239.50 20.50  243
#> 994378:     52          19534        Tarik Black  WR  IND  14 239.50 20.50  243
#>         scrape_date week projection gp_model projected_score
#>      1:  2021-09-03    1      20.62        1           20.62
#>      2:  2021-09-03    2      13.92        1           13.92
#>      3:  2021-09-03    3      19.22        1           19.22
#>      4:  2021-09-03    4      20.24        1           20.24
#>      5:  2021-09-03    5      24.22        1           24.22
#>     ---                                                     
#> 994374:  2021-09-03   10       2.10        0            0.00
#> 994375:  2021-09-03   11       0.00        0            0.00
#> 994376:  2021-09-03   12       0.00        0            0.00
#> 994377:  2021-09-03   13       0.00        0            0.00
#> 994378:  2021-09-03   14       0.00        0            0.00
# }