Download latest rankings from DynastyProcess GitHub

Fetches a copy of the latest FantasyPros redraft positional rankings data from DynastyProcess.com's data repository.

Usage

ffs_latest_rankings(type = c("draft", "week"))

Arguments

type

one of "draft" or "week" - controls whether to pull preseason or inseason rankings.

Value

a dataframe with a copy of the latest FP rankings from DynastyProcess's data repository

Details

If you have any issues with the output of this data, please open an issue in the DynastyProcess data repository.

See also

Examples

# \donttest{
try({ # try block to prevent CRAN-related issues
ffs_latest_rankings()
})
#> ── nflverse FP expert rankings ─────────────────────────────────────────────────
#> ℹ Data updated: 2024-10-03 14:21:37 UTC
#> # A tibble: 1,324 × 8
#>    player           fantasypros_id pos   team    bye   ecr    sd scrape_date
#>    <chr>            <chr>          <chr> <chr> <dbl> <dbl> <dbl> <date>     
#>  1 Brian Branch     25235          DB    DET       5   1     0   2024-09-27 
#>  2 Jessie Bates III 17544          DB    ATL      12   3     1   2024-09-27 
#>  3 Julian Love      18243          DB    SEA      10   3     0   2024-09-27 
#>  4 Derwin James Jr. 17310          DB    LAC       5   3.5   1.5 2024-09-27 
#>  5 Budda Baker      16733          DB    ARI      11   4.5   0.5 2024-09-27 
#>  6 Jaquan Brisker   23719          DB    CHI       7   8     2   2024-09-27 
#>  7 Kyle Dugger      19337          DB    NEP      14   8.5   2.5 2024-09-27 
#>  8 Kyle Hamilton    23494          DB    BAL      14   8.5   1.5 2024-09-27 
#>  9 Joshua Metellus  19580          DB    MIN       6  10.5   2.5 2024-09-27 
#> 10 Xavier McKinney  19216          DB    GBP      10  11     2   2024-09-27 
#> # ℹ 1,314 more rows
# }