Get league starter positions

This function lightly wraps ffscrapr::ff_starter_positions() and cleans up some abbreviations (PK -> K)

Usage

ffs_starter_positions(conn)

Arguments

conn

a connection object as created by ffscrapr::ff_connect() and friends.

Value

A tidy dataframe of positional lineup rules, one row per position with minimum and maximum starters as well as total starter calculations.

Examples

# \donttest{
# cached examples
try({ # try block to prevent CRAN-related issues
conn <- .ffs_cache("mfl_conn.rds")
ffs_starter_positions(conn)
})
#> # A tibble: 9 × 7
#>   pos     min   max offense_starters defense_starters kdst_starters
#>   <chr> <int> <int>            <dbl>            <dbl>         <int>
#> 1 CB        2     3               11               11             0
#> 2 DE        2     3               11               11             0
#> 3 DT        1     2               11               11             0
#> 4 LB        3     4               11               11             0
#> 5 QB        1     2               11               11             0
#> 6 RB        2     6               11               11             0
#> 7 S         2     3               11               11             0
#> 8 TE        1     5               11               11             0
#> 9 WR        3     7               11               11             0
#> # ℹ 1 more variable: total_starters <int>
# }