FP injury table

This dataframe contains a column (prob_gp) for each positional ranking that describes the probability of a player with that preseason ADP playing in a given game. It is modelled from historical rankings data and the number of games played per season for a given positional rank.

Usage

fp_injury_table

Format

An object of class tbl_df (inherits from tbl, data.frame) with 692 rows and 3 columns.