Simulate Fantasy Week

This function simulates a single upcoming week using the same methodology as in the season-long simulation, ff_simulate().

ff_simulate_week(
  conn,
  n = 1000,
  best_ball = FALSE,
  seed = NULL,
  base_seasons = 2012:2020,
  actual_schedule = TRUE,
  replacement_level = FALSE,
  pos_filter = c("QB", "RB", "WR", "TE", "K"),
  verbose = NULL,
  return = c("default", "all")
)

Arguments

conn

an connection to a league made with ff_connect() and friends (required)

n

number of times to simulate the upcoming week, default is 1000

best_ball

a logical: are weekly wins based on optimal lineups?

seed

an integer to control reproducibility

base_seasons

a numeric vector that selects seasons as base data, earliest available is 2012

actual_schedule

a logical: use actual ff_schedule? default is TRUE

replacement_level

a logical: use best available on waiver as replacement level? defaults to FALSE for upcoming week simulations

pos_filter

a character vector of positions to filter/run, default is c("QB","RB","WR","TE","K")

verbose

a logical: print status messages? default is TRUE, configure with options(ffsimulator.verbose)

return

one of c("default", "all") - what objects to return in the output list

Value

an ff_simulation object which can be passed to plot() and contains the output data from the simulation.

See also

Examples

# \donttest{
try({ # try block to prevent CRAN-related issues
conn <- mfl_connect(2021, 22627)
ff_simulate_week(conn, n = 1000, actual_schedule = TRUE)
})
#> ── Starting simulation 2021-12-20 16:12:35 ─────────────────────────────────────
#> ℹ Importing data
#> ✔ Importing data ... done
#> 
#> ℹ Generating Projections
#> ✔ Generating Projections ... done
#> 
#> ℹ Calculating Roster Scores
#> ✔ Calculating Roster Scores ... done
#> 
#> ℹ Optimizing Lineups
#> ✔ Optimizing Lineups ... done
#> 
#> ℹ Building Schedules
#> ✔ Building Schedules ... done
#> 
#> ℹ Summarising Simulation Data
#> ✔ Summarising Simulation Data ... done
#> 
#> ── Simulation complete! 2021-12-20 16:12:55 ────────────────────────────────────
#> <ff_simulation_week: 1000 simulated weeks of Four-Eight Dynasty League>
#> List of 6
#>  $ summary_simulation:Classes ‘data.table’ and 'data.frame':	4 obs. of  11 variables:
#>  $ summary_week      :Classes ‘data.table’ and 'data.frame':	4000 obs. of  16 variables:
#>  $ roster_scores     :Classes ‘data.table’ and 'data.frame':	291000 obs. of  23 variables:
#>  $ projected_scores  :Classes ‘data.table’ and 'data.frame':	291000 obs. of  13 variables:
#>  $ league_info       : tibble [1 × 17] (S3: tbl_df/tbl/data.frame)
#>  $ simulation_params :List of 7
# }