Connects ff_scoringhistory to past ADP rankings

The backbone of the ffsimulator resampling process is coming up with a population of weekly outcomes for every preseason positional rank. This function creates that dataframe by connecting historical FantasyPros.com rankings to nflfastR-based scoring data, as created by ffscrapr::ff_scoringhistory().

ffs_adp_outcomes(
  scoring_history,
  gp_model = "simple",
  pos_filter = c("QB", "RB", "WR", "TE")
)

Arguments

scoring_history

a scoring history table as created by ffscrapr::ff_scoringhistory()

gp_model

either "simple" or "none" - simple uses the average games played per season for each position/adp combination, none assumes every game is played.

pos_filter

a character vector: filter the positions returned to these specific positions, default: c("QB","RB","WR","TE)

Value

a dataframe with position, rank, probability of games played, and a corresponding nested list per row of all week score outcomes.

See also

Examples

# \donttest{
# cached data
scoring_history <- .ffs_cache("mfl_scoring_history.rds")

ffs_adp_outcomes(scoring_history, gp_model = "simple")
#> # A tidytable: 640 × 6
#>    pos    rank prob_gp week_outcomes player_name fantasypros_id
#>    <chr> <dbl>   <dbl> <list>        <list>      <list>        
#>  1 QB        1   0.919 <dbl [88]>    <chr [6]>   <chr [6]>     
#>  2 QB        2   0.919 <dbl [80]>    <chr [6]>   <chr [6]>     
#>  3 QB        3   0.919 <dbl [83]>    <chr [6]>   <chr [6]>     
#>  4 QB        4   0.919 <dbl [84]>    <chr [6]>   <chr [6]>     
#>  5 QB        5   0.919 <dbl [95]>    <chr [6]>   <chr [6]>     
#>  6 QB        6   0.918 <dbl [81]>    <chr [6]>   <chr [6]>     
#>  7 QB        7   0.917 <dbl [82]>    <chr [6]>   <chr [6]>     
#>  8 QB        8   0.915 <dbl [78]>    <chr [6]>   <chr [6]>     
#>  9 QB        9   0.913 <dbl [83]>    <chr [6]>   <chr [6]>     
#> 10 QB       10   0.911 <dbl [83]>    <chr [6]>   <chr [6]>     
#> # … with 630 more rows
ffs_adp_outcomes(scoring_history, gp_model = "none")
#> # A tidytable: 644 × 6
#>    pos    rank prob_gp week_outcomes player_name fantasypros_id
#>    <chr> <dbl>   <dbl> <list>        <list>      <list>        
#>  1 QB        1       1 <dbl [88]>    <chr [6]>   <chr [6]>     
#>  2 QB        2       1 <dbl [80]>    <chr [6]>   <chr [6]>     
#>  3 QB        3       1 <dbl [83]>    <chr [6]>   <chr [6]>     
#>  4 QB        4       1 <dbl [84]>    <chr [6]>   <chr [6]>     
#>  5 QB        5       1 <dbl [95]>    <chr [6]>   <chr [6]>     
#>  6 QB        6       1 <dbl [81]>    <chr [6]>   <chr [6]>     
#>  7 QB        7       1 <dbl [82]>    <chr [6]>   <chr [6]>     
#>  8 QB        8       1 <dbl [78]>    <chr [6]>   <chr [6]>     
#>  9 QB        9       1 <dbl [83]>    <chr [6]>   <chr [6]>     
#> 10 QB       10       1 <dbl [83]>    <chr [6]>   <chr [6]>     
#> # … with 634 more rows
# }