Generate fantasy schedules

This function generates random head to head schedules for a given number of seasons, teams, and weeks.

ffs_build_schedules(n_teams, n_seasons = 100, n_weeks = 14, seed = NULL)

Arguments

n_teams
number of teams in simulation

n_seasons
number of seasons to simulate, default = 100

n_weeks
number of weeks per season, default = 14

seed
an integer to control reproducibility

Value

a dataframe of schedules

Details

It starts with the circle method for round robin scheduling, grows or shrinks the schedule to match the required number of weeks, and then shuffles both the order that teams are assigned in and the order that weeks are generated. This doesn't "guarantee" unique schedules, but there are n_teams! x n_weeks! permutations of the schedule so it's very very likely that the schedules are unique (3x10^18 possible schedules for a 12 team league playing 13 weeks).

See also

Examples

# \donttest{
  ffs_build_schedules(n_teams = 12, n_seasons = 1, n_weeks = 14)

#> # A tibble: 168 × 4
#>    season  week  team opponent
#>     <int> <int> <int>    <int>
#>  1      1     1     1        6
#>  2      1     1     2        9
#>  3      1     1     3       11
#>  4      1     1     4        7
#>  5      1     1     5       12
#>  6      1     1     6        1
#>  7      1     1     7        4
#>  8      1     1     8       10
#>  9      1     1     9        2
#> 10      1     1    10        8
#> # … with 158 more rows

# }