Download latest rankings from DynastyProcess GitHub

Fetches a copy of the latest FantasyPros redraft positional rankings data from DynastyProcess.com's data repository.

ffs_latest_rankings()

Value

a dataframe with a copy of the latest FP rankings from DynastyProcess's data repository

Details

If you have any issues with the output of this data, please open an issue in the DynastyProcess data repository.

See also

Examples

# \donttest{

ffs_latest_rankings()

#> # A tibble: 1,301 x 9
#>    player fantasypros_id pos   team    bye   ecr    sd sportradar_id scrape_date
#>    <chr>  <chr>          <chr> <chr> <dbl> <dbl> <dbl> <chr>         <date>     
#>  1 Budda… 16733          DB    ARI      12  1.5   0.5  5ce20f3e-0f0… 2021-07-16 
#>  2 Jamal… 16375          DB    SEA       9  1.63  0.48 1a92b0ad-3eb… 2021-07-16 
#>  3 Jerem… 19429          DB    CAR      13  5.38  2.55 5f623fbc-415… 2021-07-16 
#>  4 Jorda… 11759          DB    BUF       7  5.63  2.69 95fab6fa-3ee… 2021-07-16 
#>  5 Derwi… 17310          DB    LAC       7  6     1.87 01c52412-725… 2021-07-16 
#>  6 Jessi… 17544          DB    CIN      10  7.38  2    89c84a59-18a… 2021-07-16 
#>  7 Jabri… 16382          DB    NYG      10  8.5   2.4  45ef2670-238… 2021-07-16 
#>  8 Lando… 13895          DB    WAS       9  8.75  3.03 a9c41c5b-0dc… 2021-07-16 
#>  9 John … 16747          DB    CLE      13 10     4.21 8c824157-eb3… 2021-07-16 
#> 10 Justi… 15649          DB    DEN      11 10.4   2.5  2df474e5-711… 2021-07-16 
#> # … with 1,291 more rows

# }