Optimise Lineups

Calculates optimal lineups for all franchises in the dataframe based on a table of lineup constraints.

ffs_optimise_lineups(
  roster_scores,
  lineup_constraints,
  lineup_efficiency_mean = 0.775,
  lineup_efficiency_sd = 0.05,
  best_ball = FALSE,
  parallel = FALSE
)

ffs_optimize_lineups(
  roster_scores,
  lineup_constraints,
  lineup_efficiency_mean = 0.775,
  lineup_efficiency_sd = 0.05,
  best_ball = FALSE,
  parallel = FALSE
)

Arguments

roster_scores
a dataframe as generated by ffs_score_rosters() - should contain columns like: projected_score, pos, and player_id

lineup_constraints
a dataframe as generated by ffscrapr::ff_starter_positions() - should contain columns pos, min, max, and offense_starters

lineup_efficiency_mean
the average lineup efficiency to use, defaults to 0.775

lineup_efficiency_sd
the standard deviation of lineup efficiency, defaults to 0.05

best_ball
a logical: FALSE will apply a lineup efficiency factor and TRUE uses optimal scores as actual scores, default = FALSE

parallel
a logical: TRUE will run the optimization in parallel, requires the furrr and future packages as well as setting future::plan() in advance/externally. Default FALSE.

Value

a dataframe of what each team scored for each week

Details

Lineup efficiency is the percentage of optimal/best-ball score that is used as the actual score - by default, the lineup efficiency for a team in non-best-ball settings is normally distributed around a mean of 77.5% and a standard deviation of 5%.

See also

Examples

# \donttest{
# cached examples
roster_scores <- .ffs_cache("roster_scores.rds")
lineup_constraints <- .ffs_cache("mfl_lineup_constraints.rds")

ffs_optimise_lineups(roster_scores, lineup_constraints)

#> # A tibble: 240 × 9
#>    franchise_id franchise_name                 league_id season  week optimal_score
#>    <chr>        <chr>                          <chr>      <int> <int>         <dbl>
#>  1 0003         Alberta Armadillos             22627          1     9          197.
#>  2 0010         [Redacted]                     22627          2     8          182.
#>  3 0012         Indigo Plateau Elite           22627          1    10          201.
#>  4 0009         Kill Devil Hills Carpetbaggers 22627          2     2          155.
#>  5 0004         Electric Woodies               22627          1     5          223.
#>  6 0008         Domination Station             22627          1     8          219.
#>  7 0008         Domination Station             22627          1     2          194.
#>  8 0001         Mexico City Space Force        22627          2     3          181.
#>  9 0009         Kill Devil Hills Carpetbaggers 22627          2     8          160.
#> 10 0009         Kill Devil Hills Carpetbaggers 22627          2     9          185.
#> # … with 230 more rows, and 3 more variables: optimal_lineup <list>,
#> #   lineup_efficiency <dbl>, actual_score <dbl>

# }