Summarise simulation outputs

These functions are used to summarise the simulation outputs, typically by joining the optimal scores with a matching schedule.

ffs_summarise_week(optimal_scores, schedules)

ffs_summarise_season(summary_week)

ffs_summarise_simulation(summary_season)

ffs_summarize_week(optimal_scores, schedules)

ffs_summarize_season(summary_week)

ffs_summarize_simulation(summary_season)

Arguments

optimal_scores
a dataframe of optimized lineups as created by ffs_optimize_lineups()

schedules
a dataframe of schedules as created by ffs_build_schedules()

summary_week
a dataframe as created by ffs_summarise_week()

summary_season
a dataframe as created by ffs_summarise_season()

Value

ffs_summarise_week: a dataframe summarising team results by simulation week

ffs_summarise_season: a dataframe summarising franchise results across each simulation season

ffs_summarise_simulation: a dataframe summarising franchise results across the simulation

See also

Examples

# \donttest{
# cached examples
optimal_scores <- .ffs_cache("optimal_scores.rds")
schedules <- .ffs_cache("schedules.rds")

summary_week <- ffs_summarise_week(optimal_scores, schedules)
summary_week

#> # A tibble: 240 × 15
#>    season season_week franchise_name   optimal_score lineup_efficien… team_score
#>     <dbl>       <dbl> <chr>                    <dbl>            <dbl>      <dbl>
#>  1      1           1 Mexico City Spa…          200.            0.835      167. 
#>  2      1           1 Dirty Dgens               154.            0.88       136. 
#>  3      1           1 Alberta Armadil…          126.            0.684       86.1
#>  4      1           1 Electric Woodies          144.            0.752      108. 
#>  5      1           1 Motor City Mutts          185.            0.803      148. 
#>  6      1           1 Hoth City Taunt…          110.            0.709       78.1
#>  7      1           1 Paw Patrol                166.            0.713      118. 
#>  8      1           1 Domination Stat…          197.            0.851      167. 
#>  9      1           1 Kill Devil Hill…          129.            0.811      105. 
#> 10      1           1 [Redacted]                145.            0.797      115. 
#> # … with 230 more rows, and 9 more variables: opponent_score <dbl>,
#> #   result <chr>, opponent_name <chr>, allplay_wins <dbl>, allplay_games <dbl>,
#> #   allplay_pct <dbl>, league_id <chr>, franchise_id <chr>,
#> #   optimal_lineup <list>

summary_season <- ffs_summarise_season(summary_week)
summary_season

#> # A tibble: 24 × 12
#>    season league_id franchise_id franchise_name h2h_wins h2h_winpct allplay_wins
#>     <dbl> <chr>     <chr>        <chr>             <int>      <dbl>        <dbl>
#>  1      1 22627     0001         Mexico City S…        8        0.8           75
#>  2      1 22627     0002         Dirty Dgens           1        0.1           23
#>  3      1 22627     0003         Alberta Armad…        3        0.3           47
#>  4      1 22627     0004         Electric Wood…        6        0.6           50
#>  5      1 22627     0005         Motor City Mu…        6        0.6           54
#>  6      1 22627     0006         Hoth City Tau…        4        0.4           36
#>  7      1 22627     0007         Paw Patrol            7        0.7           52
#>  8      1 22627     0008         Domination St…        5        0.5           74
#>  9      1 22627     0009         Kill Devil Hi…        4        0.4           45
#> 10      1 22627     0010         [Redacted]            5        0.5           51
#> # … with 14 more rows, and 5 more variables: allplay_games <dbl>,
#> #   allplay_winpct <dbl>, points_for <dbl>, points_against <dbl>,
#> #   potential_points <dbl>

summary_simulation <- ffs_summarise_simulation(summary_season)
summary_simulation

#> # A tibble: 12 × 11
#>    league_id franchise_id franchise_name             seasons h2h_wins h2h_winpct
#>    <chr>     <chr>        <chr>                        <int>    <dbl>      <dbl>
#>  1 22627     0012         Indigo Plateau Elite             2      7.5       0.75
#>  2 22627     0001         Mexico City Space Force          2      8         0.8 
#>  3 22627     0011         TB12                             2      5.5       0.55
#>  4 22627     0008         Domination Station               2      5         0.5 
#>  5 22627     0003         Alberta Armadillos               2      4         0.4 
#>  6 22627     0010         [Redacted]                       2      4.5       0.45
#>  7 22627     0005         Motor City Mutts                 2      5         0.5 
#>  8 22627     0009         Kill Devil Hills Carpetba…       2      5         0.5 
#>  9 22627     0004         Electric Woodies                 2      6         0.6 
#> 10 22627     0007         Paw Patrol                       2      4.5       0.45
#> 11 22627     0002         Dirty Dgens                      2      2         0.2 
#> 12 22627     0006         Hoth City Tauntauns              2      3         0.3 
#> # … with 5 more variables: allplay_wins <dbl>, allplay_winpct <dbl>,
#> #   points_for <dbl>, points_against <dbl>, potential_points <dbl>

# }